Consistently use headerStructName for peripherals' module and structure identifiers
PR #37 uses the headerStructName field to define module and structure
names. However, this change was not consistely applied throughout the
generated code -- most references to peripherals use the peripheral name
even if a header struct name is explicitly set. PSoC SVD files produce
many compile errors when #37 is applied due to this inconsistency.

This commit changes all references to peripheral structures and modules
to use the `module_id` and `struct_id` fields instead of inferring it
from the peripheral names. Additionally, the peripheral `module_id` is
now derived from the `headerStructName` for consistency with the logic
in `push_current_item_svd_path`.

I encountered some issues with feature flags -- the existing logic in
`Cargo_toml.tera` assumed that, in the case of a derived peripheral, the
peripheral's feature flag would be the peripheral's name converted to a
module identifier, and that the parent feature flag would be the
peripheral's module identifier. However, in the case of peripheral
derived from a struct with a `headerStructName`, this logic fails as the
parent peripheral's name is not the same as its module identifier. I
fixed this by changing the feature flag name to match the module name
instead of the peripheral name. This change means derived peripherals no
longer have separate feature flags from their parents.

Lastly, this commit adds a test case of a simple peripheral with a
headerStructName, which fails before this commit and passes after.