Use a custom type to represent clusters instead of an array
Introduce a `Cluster` type that represents a cluster of register blocks,
with an `Index` impl that computes the address of the register we're
interested in, instead of computing all the register addresses up front
to store them in an array. This provides a significant improvement in
code size (see #46).

We have to do some trickery in order to support subscripting, because
the `Index` trait only supports indexes returing references. This commit
implements the solution proposed by @pellico: represent a register block
within a cluster as a reference to a zero-sized type. Any non-null,
sufficiently aligned reference to a ZST is valid, so we can forge
references to ZSTs and use the address of the reference to represent the
address of the register block. This is sound as long as we never expose
an *owned* ZST to safe code, and we mark the cluster ZST types as
`#[non_exhaustive]` so users can't construct their own.